-codecs: added SID codec (C64 chiptunes player) -codecs: bugfix for wav codec -snd: some changes -libmath: added missing math_tan
git-svn-id: http://svn.code.sf.net/p/archopen/code@137 18109dc0-e51e-0410-b7de-bd5de724b689

+++ ArchOpen/trunk/medios/snd/codecs/wav/wav.c
@@ -116,7 +116,7 @@ void wav_tagRequest(char * name,TAG * tag){
                     size=wav_findChunk("data",false,f);
                     if(size>=0){
                         tag->badFile=false;
-                        tag->length=size*HZ/fmt.avgBytesPerSec;
+                        tag->length=(long long)size*HZ/fmt.avgBytesPerSec;
                     }
                 }
             }
